fix: add ltft component loading
